(#88) Rename 유저/issue88 : 구글 소셜 로그인 도입 - 폴더 구조 수정
